fix(deviations): prevent Int64 overflow in variance denominators
Ne·2·m^4 and Ne·6·m^4 silently wrapped in Int64 arithmetic at m≳1e4
(e.g. Ne=3e6, m=1e4: denom should be 6e22 but wrapped to 1.5e18, making
MDEV ~200× too large and MHDEV negative→clamped to 0). Same issue in
_simple_avar, _simple_mdev (noise_id), and htotdev/mhtotdev block-variance
denominators. Promote 2→2.0, 6→6.0, m→Float64(m) before multiplying.

Verified: MDEV/ADEV ratio now stays ~0.7 for WFM across m=1..500_000 on
N=3M synthetic; real GMR6000 file no longer has σ=0 holes, and mhdev_fit
now agrees with optimize_kf_nll on q_wpm, q_wfm, and vanishing q_rwfm.

Also adds ml/dataset/real_data_fit.jl diagnostic script.

+++ julia/src/deviations/allan.jl
@@ -43,7 +43,7 @@ function _adev_kernel(x::AbstractVector{<:Real}, m::Int, tau0::Real, x_cs::Abstr
     L = N - 2m
     L <= 0 && return (NaN, 0)
     d2 = @view(x[1+2m:end]) .- 2 .* @view(x[1+m:end-m]) .+ @view(x[1:L])
-    v  = sum(abs2, d2) / (L * 2 * m^2 * tau0^2)    # SP1065 Eq. 14
+    v  = sum(abs2, d2) / (L * 2.0 * Float64(m)^2 * tau0^2)    # SP1065 Eq. 14
     return (v, L)
 end
 
@@ -94,8 +94,10 @@ function _mdev_kernel(x::AbstractVector{<:Real}, m::Int, tau0::Real, x_cs::Abstr
     # Identity: s3 - 2s2 + s1 = x_cs[i+3m] - 3x_cs[i+2m] + 3x_cs[i+m] - x_cs[i]
     d  = @view(x_cs[1+3m:Ne+3m]) .- 3 .* @view(x_cs[1+2m:Ne+2m]) .+
          3 .* @view(x_cs[1+m:Ne+m]) .- @view(x_cs[1:Ne])
-    # Divide by 2*m^4*tau0^2 because d is Σx_i (unscaled block-sum difference)
-    v  = sum(abs2, d) / (Ne * 2 * m^4 * tau0^2)
+    # Divide by 2*m^4*tau0^2 because d is Σx_i (unscaled block-sum difference).
+    # Float64 promotion on the denominator prevents Int64 overflow for large m·Ne
+    # (e.g. Ne·2·m^4 wraps silently at m≳1e4 with Int64 arithmetic).
+    v  = sum(abs2, d) / (Ne * 2.0 * Float64(m)^4 * tau0^2)
     return (v, Ne)
 end
